Updates to make the COPPA parent email work
