+++ app/Authenticate/controllers.py
@@ -1,9 +1,7 @@
 # Import the database object from the main app module
 import logging
-import uuid
-import datetime
-
 import Failures
+
 from app import db
 from app.User import services as user_services
 from app.RateLimiting import services as rate_limiting_services
