+++ templates/en/confirm-parent/blocklyprop/header.mustache
@@ -1,1 +1,2 @@
-{{! This is the email Subject line }}New user registration confirmation request
+{{! This is the email Subject line }}
+New user registration confirmation request
