+++ templates/en/confirm-parent/blocklyprop/plain.mustache
@@ -11,6 +11,20 @@ received as the parent or guardian of the registrant. If this is incorrect,
 please accept our apologies. There is nothing more you need to do. The
 request will automatically expire.
 
+Why are we sending this? In the US, the federal Children's Online Privacy
+Protection Act, known as COPPA, requires that we communicate with a parent
+or guardian if the person registering an BlocklyProp account is under the
+age of 13. The Act allows you to decline the registration request. If you
+choose this option, we will immediately remove the registration information
+and the associated account. If you choose to confirm the the request and
+activate the account, you may close the account at any time by clicking on
+the link provided below. This action will close the account and remove any
+projects that are associated with the account.
+
+A full copy of our Child Privacy Policy is available online at:
+http://localhost:8080/blockly/child-privacy-policy
+
+
 If this is your child, please confirm the registration by copying the link below into
 your browser or by navigating to the second link and confirming the student's email address.
