+++ templates/en/confirm-teacher/blocklyprop/header.mustache
@@ -1,1 +1,2 @@
-{{! This is the email Subject line }}Please confirm your student's email address for BlocklyProp
+{{! This is the email Subject line }}
+Please confirm your student's email address for BlocklyProp
